requirements.txt
redis>=2.10.1
PyYAML>=3.11
torndb>=0.2
MySQL-python>=1.2.5
msgpack-python>=0.4.4
